rename plugin.wrapper to plugin.fnWrap
